修改：     Makefile  	修改：     camera  	新文件：   header.h  	重命名：   camera.c -> init.c  	重命名：   camera.h -> init.h  	修改：     main.c  	修改：     picture/image.bmp  	修改：     picture/image.yuv  	修改：     video_capture.c  	新文件：   video_capture.h
