+++ camera/Makefile
@@ -1,3 +1,3 @@
-SRC = main.c camera.c video_capture.c
+SRC = main.c init.c video_capture.c
 camera : $(SRC)
 	gcc -g -o camera $(SRC)
